+++ tests/nonmem/test_modelfit_results.py
@@ -15,11 +15,6 @@ def test_ofv(pheno):
     assert res.ofv == 586.27605628188053
 
 
-def test_bic(testdata):
-    model = Model.create_model(testdata / 'nonmem' / 'pheno.mod')
-    assert model.modelfit_results.bic == 756.111852398327
-
-
 def test_tool_files(pheno):
     res = pheno.modelfit_results
     names = [str(p.name) for p in res.tool_files]