+++ tests/tools/test_rankfuncs.py
@@ -1,11 +1,13 @@
+from pharmpy.modeling import load_example_model
 from pharmpy.parameter import Parameters
 from pharmpy.tools.modelsearch.rankfuncs import aic, bic, ofv
 
+pheno = load_example_model("pheno")
+
 
 class DummyResults:
-    def __init__(self, ofv=None, bic=None, parameter_estimates=None):
+    def __init__(self, ofv=None, parameter_estimates=None):
         self.ofv = ofv
-        self.bic = bic
         self.parameter_estimates = parameter_estimates
 
     def __bool__(self):
@@ -16,6 +18,8 @@ class DummyModel:
     def __init__(self, name, parameters=None, **kwargs):
         self.name = name
         self.parameters = parameters
+        self.dataset = pheno.dataset
+        self.datainfo = pheno.datainfo
         if 'no_modelfit_results' in kwargs.keys():
             self.modelfit_results = None
         else:
@@ -53,8 +57,8 @@ def test_aic():
 
 
 def test_bic():
-    run1 = DummyModel("run1", bic=0)
-    run2 = DummyModel("run2", bic=-1)
-    run3 = DummyModel("run3", bic=-14)
+    run1 = DummyModel("run1", ofv=0, parameters=Parameters([]))
+    run2 = DummyModel("run2", ofv=-1, parameters=Parameters([]))
+    run3 = DummyModel("run3", ofv=-14, parameters=Parameters([]))
     res = bic(run1, [run2, run3])
     assert [run3] == res